parte1/app/requirements.txt
mysql-connector-python
pandas
joblib
scikit-learn
fastapi
uvicorn